move filters to separate module
