+++ js/form.js
@@ -1,6 +1,6 @@
 import {setDefaultMarkerPosition, resetMapMarks} from './map.js';
 import {sendFormData} from './api.js';
-import {isEscEvent} from './util.js'
+import {isEscEvent} from './util.js';
 
 const adForm = document.querySelector('.ad-form');
 const nightPriceField = adForm.querySelector('#price');
