+++ js/map.js
@@ -1,9 +1,9 @@
 /* global L:readonly */
-/* global _:readonly */
 
 import {fillCard} from './elements-generator.js';
 import {changeFormStatus} from './form.js';
 import {getMapData} from './api.js';
+import {setupFilterHandler} from './filters.js';
 
 const DEFAULT_LAT = 35.68251;
 const DEFAULT_LNG = 139.75121;
@@ -45,7 +45,7 @@ const mainMarker = L.marker(
   },
 );
 
-const showAdsOnMap = (adsArray) => {
+const showAds = (adsArray) => {
   const adsIcon = L.icon({
     iconUrl: './img/pin.svg',
     iconSize: [32, 32],
@@ -74,21 +74,6 @@ const showAdsOnMap = (adsArray) => {
 const handleMap = () => {
 
   const ALERT_SHOW_TIME = 5000;
-  const RERENDER_DELAY = 500;
-  const HIGH_PRICE_VALUE = 50000;
-  const MIDDLE_PRICE_VALUE = 10000;
-
-  const typeFilter = document.querySelector('#housing-type');
-  const roomsFilter = document.querySelector('#housing-rooms');
-  const priceFilter = document.querySelector('#housing-price');
-  const guestFilter = document.querySelector('#housing-guests');
-  const featuresList = document.querySelector('#housing-features');
-  const wifiFilter = featuresList.querySelector('input[value=wifi ]');
-  const dishwasherFilter = featuresList.querySelector('input[value=dishwasher]');
-  const parkingFilter = featuresList.querySelector('input[value=parking]');
-  const washerFilter = featuresList.querySelector('input[value=washer]');
-  const elevatorFilter = featuresList.querySelector('input[value=elevator]');
-  const conditionerFilter = featuresList.querySelector('input[value=conditioner]');
 
   const setupAddressByMarkerOnly = () => {
     addressField.readOnly = true;
@@ -121,103 +106,13 @@ const handleMap = () => {
     }, ALERT_SHOW_TIME);
   };
 
-  const filterSimpleTextField = (array, filterName, filterField) => {
-    if (filterField.value === 'any') {
-      return array;
-    }
-    return array.filter((element) => {
-      return element.offer[filterName].toString() === filterField.value
-    })
-  };
-
-  const filterPriceField = (array, filterName, filterField) => {
-    if (filterField.value === 'any') {
-      return array;
-    }
-    return array.filter((element) => {
-      switch (filterField.value) {
-        case 'high':
-          return element.offer[filterName] >= HIGH_PRICE_VALUE;
-        case 'middle':
-          return (element.offer[filterName] >= MIDDLE_PRICE_VALUE && element.offer.price < HIGH_PRICE_VALUE);
-        case 'low':
-          return element.offer[filterName] < MIDDLE_PRICE_VALUE ;
-      }
-    })
-  };
-
-  const filterFeatures = (array, filterName, filterField) => {
-    if (!filterField.checked) {
-      return array;
-    }
-    return array.filter((element) => element.offer.features.includes(filterName))
-  };
-
-  const filtersProperties = {
-    type: {
-      filterField: typeFilter,
-      filterFunction: filterSimpleTextField,
-    },
-    price: {
-      filterField: priceFilter,
-      filterFunction: filterPriceField,
-    },
-    rooms: {
-      filterField: roomsFilter,
-      filterFunction: filterSimpleTextField,
-    },
-    guests: {
-      filterField: guestFilter,
-      filterFunction: filterSimpleTextField,
-    },
-    wifi: {
-      filterField: wifiFilter,
-      filterFunction: filterFeatures,
-    },
-    dishwasher: {
-      filterField: dishwasherFilter,
-      filterFunction: filterFeatures,
-    },
-    parking: {
-      filterField: parkingFilter,
-      filterFunction: filterFeatures,
-    },
-    washer: {
-      filterField: washerFilter,
-      filterFunction: filterFeatures,
-    },
-    elevator: {
-      filterField: elevatorFilter,
-      filterFunction: filterFeatures,
-    },
-    conditioner: {
-      filterField: conditionerFilter,
-      filterFunction: filterFeatures,
-    },
-  };
-
-  const setupFilterHandler = (adsArray, filterName) => {
-    const filterField = filtersProperties[filterName].filterField;
-
-    const onFilterChange = () => {
-      let currentAdsOnMap = [...adsArray];
-      Object.keys(filtersProperties)
-        .forEach((filter) => currentAdsOnMap = filtersProperties[filter].filterFunction(currentAdsOnMap, filter, filtersProperties[filter].filterField));
-      showAdsOnMap(currentAdsOnMap);
-    };
-    filterField.addEventListener('change', _.debounce(onFilterChange, RERENDER_DELAY));
-  };
-
   const onMapLoaded = () => {
     changeFormStatus('form_fields_enabled');
     getMapData((adsArray) => {
       downloadedAds = [...adsArray];
-      showAdsOnMap(adsArray);
+      showAds(adsArray);
       changeFormStatus('filters_enabled');
-
-      for (let filterName in filtersProperties) {
-        setupFilterHandler(adsArray, filterName)
-      }
+      setupFilterHandler(adsArray);
     },
     () => showMapAlert('Не удалось загрузить объявления с сервера'),
     );
@@ -235,6 +130,6 @@ const setDefaultMarkerPosition = () => {
   addressField.value = `${DEFAULT_LAT}, ${DEFAULT_LNG}`;
 };
 
-const resetMapMarks = () => showAdsOnMap(downloadedAds);
+const resetMapMarks = () => showAds(downloadedAds);
 
-export {handleMap, setDefaultMarkerPosition, resetMapMarks};
+export {handleMap, setDefaultMarkerPosition, resetMapMarks, showAds};
